python_jarvis/requirements.txt
speech_recognition
pyttsx3
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
selenium
webdriver-manager
pywhatkit
python-dotenv